refactor: Split External Rescue components
