+++ src/index.tsx
@@ -40,8 +40,8 @@ import Privacy from "./pages/Privacy";
 import RefundRescue from "./pages/RefundRescue";
 import Rescue from "./pages/Rescue";
 import RescueEvm from "./pages/RescueEvm";
-import RescueExternal from "./pages/RescueExternal";
 import Terms from "./pages/Terms";
+import RescueExternal from "./pages/external-rescue/RescueExternal";
 import Btcpay from "./pages/products/Btcpay";
 import Client from "./pages/products/Client";
 import Pro from "./pages/products/Pro";
