+++ src/style/index.scss
@@ -1523,61 +1523,6 @@ header {
     width: 100%;
 }
 
-.mnemonic-input-container {
-    display: flex;
-    flex-direction: column;
-    margin-top: 0.625rem;
-    gap: 10px;
-
-    .mnemonic-inputs {
-        display: grid;
-        grid-template-columns: repeat(2, 1fr);
-        grid-template-rows: repeat(6, 1fr);
-        grid-auto-flow: column;
-        column-gap: 1rem;
-        row-gap: 0.5rem;
-
-        max-width: 18rem;
-        padding: 1rem;
-        margin: 0 auto;
-
-        color: var(--color-secondary-100);
-        background: var(--color-secondary-800);
-        border: 1px solid var(--color-white-15);
-        border-radius: var(--border-radius-md);
-
-        .input-box {
-            display: flex;
-            gap: 0.8rem;
-            justify-content: space-around;
-
-            .mnemonic-input {
-                padding: 0;
-
-                color: var(--color-secondary-100);
-                background: transparent;
-                font-weight: 600;
-                text-align: left;
-
-                border: none;
-                border-bottom: 1px solid var(--color-white-15);
-                border-radius: 0;
-            }
-            .mnemonic-input:focus {
-                border-bottom: 1px solid var(--color-primary);
-
-                &.invalid {
-                    border-bottom: 1px solid var(--color-error-500);
-                }
-            }
-
-            .invalid {
-                border-bottom: 1px solid var(--color-error-500);
-            }
-        }
-    }
-}
-
 .mnemonic-backup-verify-container {
     display: flex;
     flex-direction: column;
