+++ src/style/rescueExternal.scss
@@ -165,6 +165,63 @@
         display: contents;
     }
 
+    .mnemonic-input-container {
+        display: flex;
+        flex-direction: column;
+        margin-top: 0.625rem;
+        gap: 10px;
+
+        .mnemonic-inputs {
+            display: grid;
+            grid-template-columns: repeat(2, 1fr);
+            grid-template-rows: repeat(6, 1fr);
+            grid-auto-flow: column;
+            column-gap: 1rem;
+            row-gap: 0.5rem;
+
+            max-width: 19rem;
+            padding: 1rem;
+            margin: 0 auto;
+
+            color: var(--color-secondary-100);
+            background: var(--color-secondary-800);
+            border: 1px solid var(--color-white-15);
+            border-radius: var(--border-radius-md);
+
+            .input-box {
+                display: flex;
+                gap: 0.8rem;
+                justify-content: space-around;
+            }
+
+            .mnemonic-input {
+                min-height: 0;
+                height: 1.4rem;
+                padding: 0;
+
+                color: var(--color-secondary-100);
+                background: transparent;
+                font-weight: 600;
+                text-align: left;
+
+                border: none;
+                border-bottom: 1px solid var(--color-white-15);
+                border-radius: 0;
+            }
+            .mnemonic-input:focus {
+                border-bottom: 1px solid var(--color-primary);
+
+                &.invalid {
+                    border-bottom: 1px solid var(--color-error-500);
+                }
+            }
+
+            .invalid {
+                border-bottom: 1px solid var(--color-error-500);
+            }
+        }
+    }
+
     .rescue-external-wallet-slot .btn::before {
         content: "";
         display: inline-block;
